requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
python-telegram-bot
feedparser
requests
